Updated ent_tank so that it's ready to be made into a class
